Update Search Input
Refactor search input dropdown structure:
Previously, the search input dropdown would always render as a `<ul>`, even when there were no options.
This resulted in invalid HTML, because there would be `<divs>` as direct children of `<ul>`.

Refactored component to render the box as a `<div>`, and only render a Combobox.Options component
when there are available options.

Other changes:

- Update label text

- Add placeholder

- Update colors

- Add focus ring